requirements.txt
discord.py==1.7.1
pdf2image==1.16.0